Rename (ZK Login) Signature to ZKLoginSignature
