+++ Assets/Sui-Unity-SDK/Code/Sui.ZKLogin/ZKLoginSignature.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 21defd90e4a33403d8607f823c08ae2b
+guid: 861913107c6e549769e04494584c680f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
